// src/app/models/Gallery.ts
export class Gallery{
    GalleryID: number;
    ImageURL: string;
    OrignalName: string;
    EncryptedName: string; 
    GalleryTypeID: number;
    ObjectID: number;
    EmployeeID: number;
    JobTaskID: number;
    CreatedOn: boolean;
    CreatedBy: number;
    ModifiedBy: number;
    IsDeleted: boolean;
    ImageDataUrl: string;
    IsXLImage:  boolean;
    IsVideo:  boolean;
    IsModified : boolean;
}